+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - Fix up duplicate webhook triggering on manual event update.
 - Add support for scheduled event injection at run-time.
 - Add support for self-contained Docker image and Kubernetes deployment.
+- Add support for prometheus metrics
 
 ## 1.1.1